+++ src/context/useShoppingContext.test.tsx
@@ -3,14 +3,13 @@ vi.mock("../../api/cartItem");
 
 import { render, screen, waitFor } from "@testing-library/react";
 import { vi } from "vitest";
-import * as productApi from "../../api/product";
-import * as cartApi from "../../api/cartItem";
-import ShopPage from "../../page/ShopPage";
-import { ContextProvider } from "./useShoppingContext";
-
-import { mockCartItemResponse } from "../../mock/CartProducts";
-import { mockProductResponse } from "../../mock/Products";
-import CartProductContainer from "../../component/feature/CartProductContainer/CartProductContainer";
+import * as productApi from "../api/product";
+import * as cartApi from "../api/cartItem";
+import ShopPage from "../page/ShopPage";
+import { mockCartItemResponse } from "../mock/CartProducts";
+import { mockProductResponse } from "../mock/Products";
+import CartProductContainer from "../component/feature/CartProductContainer/CartProductContainer";
+import { ContextProvider } from "./ContextProvider";
 
 describe("useContext 훅을 이용해서 productList를 관리한다.", () => {
   beforeEach(() => {
